APIS-4741 - Split DocumentationController  - API dynamic content lives in ApiDocumentationController
