+++ app/uk/gov/hmrc/apidocumentation/views/ApisFilteredView.scala.html
@@ -27,7 +27,7 @@
 
     <h1 class="heading-xlarge">Filtered API documentation</h1>
 
-    <a class="bold-small" href="@controllers.routes.DocumentationController.apiIndexPage(None, None, None).url">View the full API list </a>
+    <a class="bold-small" href="@controllers.routes.ApiDocumentationController.apiIndexPage(None, None, None).url">View the full API list </a>
 
     @documentFilter(documentsByCategory, filter)
