+++ app/uk/gov/hmrc/apidocumentation/views/AuthorisationAppRestrictedEndpointsView.scala.html
@@ -61,7 +61,7 @@
             If the endpoint requires a <a href="@controllers.routes.DocumentationController.referenceGuidePage()#scopes">scope</a>,
             your application must include this scope when creating the access token.
         </p>
-        <p>Read the <a href="@controllers.routes.DocumentationController.apiIndexPage(None, None, None)">API documentation</a> for authorisation rules for specific API endpoints.</p>
+        <p>Read the <a href="@controllers.routes.ApiDocumentationController.apiIndexPage(None, None, None)">API documentation</a> for authorisation rules for specific API endpoints.</p>
 
         <div class="panel panel-border-wide">
             <p>The access token lasts for 4 hours. When it expires you must generate a new token.</p>
