+++ app/uk/gov/hmrc/apidocumentation/views/AuthorisationUserRestrictedEndpointsView.scala.html
@@ -60,7 +60,7 @@
 
     <p>There are separate Government Gateway accounts for individuals, agents and organisations. For individuals and organisations, the access token only gives access to the end user’s own data. For agents the access token gives access to their clients’ data.</p>
 
-    <p>Authorisation rules for specific API endpoints are given in the <a href="@controllers.routes.DocumentationController.apiIndexPage(None, None, None)">API documentation</a>.</p>
+    <p>Authorisation rules for specific API endpoints are given in the <a href="@controllers.routes.ApiDocumentationController.apiIndexPage(None, None, None)">API documentation</a>.</p>
 
     <p>The access token lasts for 4 hours. When it expires, it can be refreshed using a single-use refresh token. After 18 months you can no longer refresh the access token and the end user must grant authority again.</p>
