+++ app/uk/gov/hmrc/apidocumentation/views/FraudPreventionView.scala.html
@@ -48,7 +48,7 @@
     <p>Supplying header information for all our APIs will become mandatory - so we recommend designing it into your applications now.</p>
     <p>It is mandatory to provide header information when using the following APIs:</p>
     <ul class="list list-bullet">
-      <li><a href="@controllers.routes.DocumentationController.renderApiDocumentation("vat-api", "1.0", None).url">VAT (MTD)</a> from 1st April 2019</li>
+      <li><a href="@controllers.routes.ApiDocumentationController.renderApiDocumentation("vat-api", "1.0", None).url">VAT (MTD)</a> from 1st April 2019</li>
     </ul>
     <p>We have published our <a href='@routes.Assets.versioned("content/documentation/TxM DPIA v3 1 Public.pdf")'>TxM Data Protection Impact Assessment</a> (DPIA) in line with best practice guidance from the Information Commissioner’s Office. This is to:</p>
     <ul class="list list-bullet">
