+++ app/uk/gov/hmrc/apidocumentation/views/IndexView.scala.html
@@ -74,7 +74,7 @@
             <section class="section section--branded">
                 <h2>Explore APIs</h2>
                 <ul>
-                    <li><a href="@controllers.routes.DocumentationController.apiIndexPage(None, None, None).url">API documentation</a></li>
+                    <li><a href="@controllers.routes.ApiDocumentationController.apiIndexPage(None, None, None).url">API documentation</a></li>
                     <li><a href="https://www.gov.uk/topic/dealing-with-hmrc/software-development">XML APIs</a></li>
                     <li><a href="@controllers.routes.DocumentationController.referenceGuidePage()">Reference guide</a></li>
                     <li><a href="@controllers.routes.DocumentationController.developmentPracticesPage()">Development practices</a></li>
