+++ app/uk/gov/hmrc/apidocumentation/views/ReferenceView.scala.html
@@ -541,7 +541,7 @@
 
         <p>These are translated to human-readable descriptions that are shown to the user before they grant authority. This makes sure the user understands and gives access to your application.</p>
 
-        <p>The scope for each user-restricted endpoint is defined in the <a href="@controllers.routes.DocumentationController.apiIndexPage(None, None, None).url">API documentation</a>.</p>
+        <p>The scope for each user-restricted endpoint is defined in the <a href="@controllers.routes.ApiDocumentationController.apiIndexPage(None, None, None).url">API documentation</a>.</p>
 
         <p>For details about OAuth 2.0 and scopes, see <a href="@controllers.routes.AuthorisationController.authorisationUserRestrictedEndpointsPage().url">authorisation</a>.</p>
