+++ app/uk/gov/hmrc/apidocumentation/views/TutorialsView.scala.html
@@ -383,6 +383,6 @@ Long expiresIn = response.getExpiresIn();</pre>
 <br/>
 
 <p>You should now be ready to start integrating with <a
-        href="@controllers.routes.DocumentationController.apiIndexPage(None, None, None).url">our APIs</a>.</p>
+        href="@controllers.routes.ApiDocumentationController.apiIndexPage(None, None, None).url">our APIs</a>.</p>
 
 }
